Time, PHYSFS music
- SDL_GetTicks and time()/localtime() replaced by Time::clock(), Time::local()
- all times and tick counters are uint
- oz
  * Time class
  * defines for Win32 headers moved to internal header winedefs.h
- client
  * music works through PHYSFS

+++ src/client/ui/InfoFrame.cc
@@ -74,7 +74,7 @@ void InfoFrame::onDraw()
     return;
   }
 
-  if( lastId != camera.tagged || uint( timer.ticks - lastTicks ) >= uint( REFRESH_INTERVAL ) ) {
+  if( lastId != camera.tagged || timer.ticks - lastTicks >= REFRESH_INTERVAL ) {
     const Bot* tagged = static_cast<const Bot*>( camera.taggedObj );
 
     if( tagged->flags & Object::BOT_BIT ) {
